refactor: move the CBlock claim accessors into gridcoin/block_claims.cpp (#3125 C9)
Move CBlock::GetClaim / PullClaim / GetSuperblock (both overloads) /
GetMint / GetMRCFees and the free GetClaimByIndex out of main.cpp into a
new gridcoin/block_claims.cpp. The member declarations already live in
primitives/block.h, but the bodies cannot go to primitives/block.cpp:
they drag GRC::Claim / GRC::MRC / CTxDB into the primitives layer.

GetClaimByIndex canonical declaration (and a GRC::ClaimOption typedef
identical to the main.h compatibility one) move to gridcoin/claim.h,
which gains the sync.h + extern-cs_main idiom for the lock annotation;
main.h keeps a verbatim compatibility duplicate in its fenced block
(claim.h is not part of the re-include set).

Pure code motion; no behavior change. block_claims.cpp registered in the
core source list.
